refactor(oxp_tty,oxp_hid): move ButtonId to oxp_tty module and deduplicate
Move ButtonId definition from serial_report.rs to oxp_tty/mod.rs as a
shared public type. Update oxp_tty/driver.rs and oxp_hid/hid_report.rs
to import from the module root. Also downgrade device open log to debug.
