+++ src/drivers/oxp_tty/mod.rs
@@ -2,6 +2,35 @@ pub mod driver;
 pub mod event;
 pub mod serial_report;
 
+use packed_struct::prelude::*;
+
+/// Button IDs shared by OXP HID and TTY drivers.
+#[derive(PrimitiveEnum_u8, Clone, Copy, PartialEq, Debug, Default)]
+pub enum ButtonId {
+    #[default]
+    None = 0x00,
+    A = 0x01,
+    B = 0x02,
+    X = 0x03,
+    Y = 0x04,
+    LeftBumper = 0x05,
+    RightBumper = 0x06,
+    LeftTrigger = 0x07,
+    RightTrigger = 0x08,
+    Menu = 0x09,
+    View = 0x0a,
+    LeftStick = 0x0b,
+    RightStick = 0x0c,
+    DpadUp = 0x0d,
+    DpadDown = 0x0e,
+    DpadLeft = 0x0f,
+    DpadRight = 0x10,
+    Guide = 0x21,
+    M1 = 0x22,
+    M2 = 0x23,
+    Keyboard = 0x24,
+}
+
 const TAKEOVER_COMMAND: [u8; 64] = [
     0x00, 0xA1, 0x3F, 0x01, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
